Attempt to clone and append to routes
This works, but with a caveat: routes added by tests persist across all
subsequent tests. See the `n_#{n}` routes.

I tried to ensure all objects were being duplicated, but some still
remain shared. We're getting into the weeds of the internals of Rails
routing and it feels very fragile; it's not a viable long-term solution.